requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Message Queue
redis==5.0.1
celery==5.3.4

# Network Protocols
pysnmp==4.4.12
ncclient==0.6.15
requests==2.31.0
httpx==0.25.2

# Monitoring & Metrics
prometheus-client==0.19.0
prometheus-api-client==0.5.3

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
